Compile plugins defined in a package when building the package (#4080)
This improves package authoring by surfacing errors before having to run a command.

Incremental builds don't load the package graph if the package structure and manifest contents haven't changed, so to make this happen we need to also store basic information about the plugins in the BuildDescription.  We don't want to incorporate these commands into the build graph at this point, since it would be a much bigger change and also complicate the compile-on-demand behavior of the plugins, plus require all clients of libSwiftPM to add a dependency on llbuild — but it's possible that we'll want to make that restructuring in the future (in particular if it would allow a plugin to use a library in an easier way).

Specific changes in this PR:

- move PluginCompilationResult from DefaultPluginScriptRunner to PluginInvocation so it can be referenced by the PluginScriptRunner protocol
- add a method to PluginScriptRunner to compile but not run a plugin and just return a compilation result
- add a PluginDescription to BuildPlan, and serialize it into the BuildDescription so we don't have to reload the graph
- add a PlugInScriptRunner to BuildOperation, and have SwiftTool (and specializations of it) pass this down to the operation when creating it
- replace the `buildToolPluginInvoker` closure with a PluginScriptRunner and pluginWorkingDirectory pair (a bit less of a layering violation)
- extend BuildOperationBuildSystemDelegateHandler to also be called when a preparation step starts, emits output, and finishes
- compile the plugins in the build set before starting the build

As a bonus this also makes build tool plugins available in all build operations, including the API Digester etc.