Disable a flaky part of the unit test `PackageToolTests.testPluginCompilationBeforeBuilding`.  Filed https://bugs.swift.org/browse/SR-15831 to investigate and fix it.
We don't want to skip the entire test since there are other good things being tested here that are not flaky, such as some of the other output as well as the exit code of the subprocess.

This also fixes a separate mistake in the test, which is not relevant to this flakiness but may cause confusion because the terminology in a fixture is wrong.