Cancelling `swift package` with Control-C doesn't stop any running plugins
Add a registry that keeps track of running plugin processes, and an API to cancel them.  We can't use `ProcessSet()` since it's tied to ToolsSupportCore's implementation of a subprocess and doesn't use Foundation.  In this case we don't need such a general solution in any case.

Added a unit test that goes to some length to try to determine that process is really gone after cancelling.

rdar://87280309