Add -dead_strip / --gc-sections for release builds (#4135)
motivation: In the past it was unsafe to pass -dead_strip because there was some
swift metadata that could be stripped. It seems that at this point all
of those cases are either fixed or marked as @llvm.used, so passing
these flags should safely reduce binary size in some cases. If there are
other cases where this isn't safe we should likely annotate them as
@llvm.used anyways.

Related: https://bugs.swift.org/browse/SR-521
https://github.com/apple/swift-package-manager/pull/215

changes: 
* Add -dead_strip / --gc-sections for release builds
* Add --disable-dead-strip flag
* Add CHANGELOG entry