Build: adjust the BuildSystemDelegate hander
llbuild does not make guarantees that the output from subcommands are
emitted in complete buffers.  This shows up particularly well on
Windows.  clang will emit diagnostics which get read in different
chunks.  By emitting these eagerly we would print partial messages which
would leave the terminal in the improper state due to the message not
being complete.  With multiple parallel threads writing the messages
partially, we would get splits in the message that would yield
unreadable messages.

Switch to a buffering approach, buffering the entire content until the
command exits.  At that time we can emit the complete message in a
"single" write (which is locked).  This ensures that the output is no
longer partially emitted and corrects the rendering with the animated
emission on Windows.

Co-authored-by: tomer doron <tomer@apple.com>