Re-add class name to '__allTests' for test discovery (#4226)
Motivation:

In https://github.com/apple/swift-package-manager/pull/3992 the class
name was stripped from the static array of tests (i.e.
`__allTests_CLASSNAME` became `__allTests`). The class name ensured the
variable was uniquely named. Without it, in a test class which is
derived from another, it is not unique and results in compile-time
errors.

Modifications:

- Reintroduce the class name
- Add tests

Result:

Dervied test classes do not result in compilation errors on Linux.