Allow binary targets to change checksums and URLs at the same time
Previously if you had a binary target, that you replaced the checksum
and URL for at the same time, SwiftPM would reject it because the
checksum had changed. In my experience this is the common case when you
are copy pasting the inclusion instructions from documentation for an
update and then you run `swift build`.

In the case of package resolution the addedOrUpdatedPackages list is
always empty.