Remove use of `AbsolutePath.appending(RelativePath(_:))`
This replaces the use of `AbsolutePath.appending(RelativePath(_:))` in
the code base in favour of `AbsolutPath(_:relativeTo:)`.  This form
allows us to compute the absolute path prior to the normalization of the
path which ensures that we can validate the path properly.  This is
particular important to enable the use of relative paths on Windows
which cannot normalize at the parent path without a reference as there
may be substitutions which need to be applied.