Very rarely the last stderr output from a plugin subprocess could get lost
This could happen because the Process termination handler only called `availableData` and not `readToEnd()`, so based on timing, there might be data that had been emitted by the plugin subprocess before it exited but that hadn't yet been received.

This caused sporadic failures of the `testLocalAndRemoteToolDependencies` unit test.

The fix is to switch to using `readToEnd()`.  This was not a problem for messages from the plugin, because they were already being read (and not collected using `availableData`).

rdar://89490777