Repair #4258
The original change was the most expedient change to get the CI
coverage restored.  After further investigation, it seems that we are
able to reconstruct the XCTest path through a public variable that has
appeared after the original Windows support was added.  This now allows
us to restore the configuration to its previous state of `private`.