PackageModel: adjust toolchain layout for Windows
Add in a migration path for adjusting the layout of the toolchain.  This
is required in order to support more than one architecture SDK to be
installed simultaneously.  The problem is primarily on the XCTest side,
where the binary is not slotted to an architecture.  Having multiple
variants of XCTest is difficult with the toolchain layout.  The XCTest
installation is located at
`%DEVELOPER_DIR%\Platforms\<Platform>.platform\Library\XCTest-<version>`.
The runtime component is located under, and the DLL is always named
`XCTest.dll`.  The following options were considered:

1. renaming `XCTest.dll` to:
  - `XCTest32.dll` (x86)
  - `XCTest64.dll` (x64)
  - `XCTest32a.dll` (ARMv7)
  - `XCTest64a.dll` (ARM64)

2. renaming `bin` to:
  - `bin32` (x86)
  - `bin64` (x64)
  - `bin32a` (ARMv7)
  - `bin64a` (ARM64)

3. adding a subdirectory to `bin`

Option 2 was deemed to be the best of the terrible options.  It would
allow a migration path (as the DLL name is embedded into the import
library and subsequently the executable).

We simultaneously update the location for the import library.  Although
it is possible to build a multi-architecture import library, we would
need generate that manually before the installer package is built and
then only install the runtime.  It seems that moving the directory would
be easier and ensure that we can keep them somewhat independent.