PackageModel: correct and improve XCTest location on Windows
This corrects a long standing issue in locating XCTest on Windows.  We
erroneously assumed that we could use `DEVELOPER_DIR` to construct a
path to XCTest.  However, this is not correct for two reasons:

  1. We may be targeting something other than Windows
  2. We may have passed `-sdk` explicitly, which would move the
     location of the SDK to something that no longer is relative
     to `DEVELOPER_DIR`.

Instead, we now construct the path relative to `SDKROOT` OR the explicit
SDK specified.  This reduces the dependency on the environment, and
makes `SDKROOT` behave as intended as well as completely drops the need
for `DEVELOPER_DIR` within SPM.

This will also restore the ability for us to have relocatable SDK
installations on Windows which eases some local testing for parallel SDK
installations.