Pass `-enable-regex-literals` to Swift compiler for packages with tools version ≥ 5.7
This is made conditional on the tools version to make sure that existing packages do not break, since the spelling of regular expressions that this flag enables could cause existing source code to be misinterpreted.  This change will also avoid passing the flag if the compiler doesn't support it.

rdar://91118955