PackageModel: support the unified swift modules for XCTest
`XCTest` accidentally did not use the directory form of the swiftmodule,
which did not matter with a single architecture being available.
Migrate to the directory form to match the SDK setup.  With the SDK
migrating to the architecture independent directory for the modules,
follow suit and provide a compatibility path that will allow use of
either.