PackageLoading: support non-Unix platforms for PkgConfig
pkg-config is a common utility on UNIX systems, but does not play very
well with Windows.  This adjusts the path separator character to match
the platform as well as uses `AbsolutePath` APIs to work with paths as
a means of getting the platform specific path separator character.  It
also introduces the ability to specify the pkgconfig tool itself which
is needed for Windows which does not support an arbitrary executable.