Basics: add a temporary workaround for Windows for archives (#5521)
The archive support currently assumes that it can invoke random tools
from the system.  There is no `unzip` available on Windows by default,
though we might be able to get away with the fact that there exists a
bundled copy in git.  However, that requires that we alter the search
path for the invocation which would get complicated.  Instead, we can
use a horrible hack of using `tar` to extract the archive given thats
built with libarchive and will simply use file detection and extracts
it.

Longer term it would be ideal to avoid the utility invocation and use
directly a library to extract archives.