Tests: adjust VFS tests for Windows
Windows requires that we have the destination exist before we create the
file to disambiguate between the file and directory symbolic link case.
Swap the ordering of the fixture creation to accommodate this.

The `isExecutable` check on Windows is more restrictive, identifying
only binary executables rather than scripts.  As a result, we must
disable one of the checks.