PackageLoading: further portability changes for pkg-config (#5527)
Windows and subsequently `pkg-config` (and `pkgconf`) use `;` as the
path delimiter rather than `:`.  Adjust the codepaths to account for
the difference in the sigil.  Additionally, because `\` is the preferred
path separator on windows, the pkg-config specification behaves slightly
differently for the treatment of `\`.  Unless we are escaping spaces or
the `\` literal, emit the `\` as a character rather than treat it as an
escape modifier.  This gives us the desired behaviour, even though it is
a slight bit more clever than the reference implementations.