Tests: adjust repository test and disable on Windows (#5543)
This adds support for Windows to the compressed repository download
test.  Unfortunately, the repository contained within the tarball
contains files which contain invalid characters for some file systems.

For the time being, disable the test, though it would be preferable to
limit the repository contents to that which can be represented on all
file systems (perhaps consider the FAT32 restrictions which likely can
cover the file systems that we are likely to run on).