Use llbuild's new `content-exclusion-patterns` node attribute to prevent it from descending into `.build` and `.git` directories (#5594)
This requires the llbuild commit a1adaff2a405c3a74925e60356431050cb9a3a9d, but it does not cause problems if the llbuild being used doesn't yet support this property.  Instead it will be ignored.

This adds a dedicated unit test with a flat package.  We should consider whether any customized scratch directory should also be included in the list.  This is a little tricky since the exclusion patterns are applied to all subdirectory name, not just those at the top level.  In testing, it doesn't actually seem to be required, which seems a little surprising.  The fix in this commit is safe and good on its own, and we can consider adding customized scratch paths later once we've looked deeper into that issue.

rdar://93982172
(cherry picked from commit fb032acf252a36917df3f1d780b78195d570ed8b)