Fix MacCatalyst minimum deployment fallback
If a package does not specify a custom deployment target for MacCatalyst, we are supposed to fallback to the iOS one, if declared. This currently does not work for test targets, because we prefer the minimum XCTest deployment target, even if it may be lower than the customized iOS one.