Validation of package products should check that a library product doesn't try to include executable targets (#5625)
Currently a library product that includes an executable target in its `targets` array only results in a warning.  This was the best that could be done before `executableTarget` was introduced, since further checking was needed to see if a `target` was a library or executable target.  Since we have explicit types now, we should promote this to an error since people are running into it and it's not clear what the problem is.

This check is limited to tools version 5.7 or later.  This change includes a new test fixture and unit test.

rdar://95782540