[5.7] SwiftPM does not build all dependency executables before invoking plugin build command (#5636)
Make the list of input dependencies of an llbuild command that runs a plugin-provided build tool also have dependencies on the paths of all the tools on which the plugin has declared a dependency, not just the one that appears as an executable in the build command.

Also adds a mildly elaborate unit test to make sure this works.

rdar://93679015
(cherry picked from commit a6ac80f6d2ecda7b7933e6f2a26656740d24a05f)