Build: initial pass to support static archives on Windows (#5720)
Introduce a SPM controlled build rule for building static libraries.
This is the intended way to use llbuild to drive the generation of
static libraries.  We would previously rely on the static default
rule intended for testing to generate the static libraries.  Not only
did this tool not properly support Windows, it would actually cause
problems on macOS due to the use of `ar` for the creation of the library
over the preferred tool - `libtool`.  We now locally determine the
correct rule and generate the command.

This is incomplete support for Windows and in fact regresses
functionality.  We no longer honour `AR` as an environment variable on
Windows and thus cannot switch the implementation of the librarian.  We
now drive the archiving through `lld-link` unconditionally while we
should prefer `link` unless otherwise requested.  This is covered as
an issue in #5719.