Fix mangled output when compiling plugins before a build starts
The output was just emitting to the output stream, but should have been going through the progress tracker and updating it the same as for any build command.

# Changes
- the callbacks that emit "build preparation step" information now uses the same queue as build messages, and tell the task tracker
- detailed preparation step output is now only emitted if a verbose log level is used
- the build operation now uses the callbacks to drive emitting plugin compilation messages

An even better approach in the future would probably be to have the plugin compilation commands be registered as llbuild tasks, which would allow them to go through the regular dependency machinery.  This would still have to happen as a separate build step before the regular build, since prebuild commands etc have to run before the build and can contribute to target sources discovery etc (i.e. it can't be part of a regular build).

rdar://98367242