Basics: migrate `TSCUtility.SQLite` to `Basics`
This moves the SQLite API wrapper to SPM from tools-support-core.  There
is currently a single user of this API, which is SPM.  By migrating the
API to SPM we can deprecate the TSC interfaces to prevent new adoption.
It also reduces the dependencies on `TSCUtility` for SPM which works
towards the longer term goal of shedding the tools-support-core
dependency.