Swift package manifest generation should not write out absolute paths for path-based package dependencies
## Motivation
path-based references were being written out as absolute during manifest source generation, even though relative is almost always (if not always) the right thing to write out in order to avoid creating packages that only work on the local machine.

## Modifications
- add a `packageDirectory` parameter to `generateManifestFileContents()` to specify where the package containing the manifest is located
- use this directory to make path-based dependencies relative
- remove the old compatibility wrapper `generatedManifestFileContents` which no longer makes sense (clients can easily switch to the regular function)
- adjust and extend unit test that covers manifest source generation for package dependencies
- in particular, that test is now more realistic to the kinds of packages that are actually useful to write

I considered also keeping a way to write out absolute path-based dependencies, but since resulting packages are inherently non-portable there didn't seem to be a reason to keep this.  It makes the API more complicated but we can add that if it's needed.

## Result
- package manifests written by the manifest source generation have absolute paths

rdar://100367134