Fix use of `extraCPPFlags`
We should not use `extraCPPFlags` as a vehicle for specifying linker flags, instead just decide on which C++ library to link based on the target triple in `BuildPlan`.

Separately, it seems worthwhile to look into refactoring `Destination` to actually allow specifying linker flags since that seems to have been an accidental feature of `extraCPPFlags` (as long as one didn't pass compiler flags there as well).

rdar://100575898