Make the code that deals with the remaining plugin output data more robust
This code in DefaultPluginScriptRunner uses NSFileHandle's `readabilityHandler`, and there were some circumstances in which this code could crash.

This change introduces local variables that keep both the handle and the handler alive long enough to not crash rather than doing it all in one expression. Note that this isn't a race condition — the handler does the correct locking. But the object lifetime was a bit ambiguous with the existing line of code, and the new one removes the ambiguity.

No new unit tests — 21 of the existing unit tests cover this, and that's how the ambiguity was found.

rdar://101229820