Tests: repair `testPackageDependencyVariations` on Windows (#5825)
The absolute path was not absolute, but when constructing an
`AbsolutePath` we would use the current drive as the path is
current drive relative.  In debug builds, this would trigger
an assertion failure in swift-tools-support-core.  Adjusting
the path representation to be OS dependent fixes the assert.

Additionally, adjust the path representation that is used to
compare the string literals. Because the path representation
is converted to a native representation, we must expectation
values accordingly.