Introduce `extraCXXFlags`, deprecate `extraCPPFlags` (#5827)
Usually, CPP is used as an abbreviation for the C preprocessor, while CXX more commonly refers to the C++ compiler. We also use `cxxSettings` in `Package.swift` manifest: https://developer.apple.com/documentation/packagedescription/target/cxxsettings.

For this wider consistency, `extraCXXFlags` is introduced as a replacement for `extraCPPFlags`. The latter is then deprecated and is made a computed property that just returns the value of `extraCXXFlags`.