Avoid redundant checks in cycle detection
Once all of a Manifests outbound edges have been traversed for cycles, it's not possible for it to participate in a cycle. So if a fully-explored Manifest is encountered again in a cycle check, exit early.

Without such an early exit, cycle detection is quadratic. This change makes it linear.

Convert tabs to spaces

Fix indentation

Add performance test for Workspace package loading

Don't disable error propagation

Move test to PackageGraph and remove fixtures