Use repository cache without update if offline (#5819)
If there's an error populating the cache, check whether we're offline and have an existing valid repository state. If yes, use cached state and emit a warning.

On macOS, we're using network reachability to determine the connection status, on other platforms we rely on crude parsing of the git error message for now.