Improve `diagnose-api-breaking-changes` error reporting
It looks like `swift-api-digester` is always returning 0, so failures to parse modules can turn into false negative breakage reports. Do a minimal sniff test of the produced JSON to see if it contains symbols and if it does not, we'll emit the full output of `swift-api-digester` which will hopefully contain any errors encountered.