Default target platform to current host version if unspecified
Previously if you created a new basic Swift package and added a
dependency, such as swift-syntax, that requires a minimum version
greater than macOS 10.13, you would have to set the platform version to
something above that in the Package.swift. With this change if the
supported platform version is unspecified, the current host version is
used instead. This makes it easier to test out projects where the
minimum version isn't important. The user can still get a hermetic
minimum version by setting whatever value they need.

https://github.com/apple/swift-syntax/blob/ef4605bc20a17f73939623fef9cec75fa89928b1/Package.swift#L28