PackageModel: rename "CC SDKs" to "destinations" (#5905)
"Cross-compilation SDK" is a bit of a mouthful to say, and we've previously denoted this as a "destination". After some consideration, I think calling "cross-compilation SDKs" as "destinations" simplifies a lot of things. Additionally, "target triple" and "target platform" are terms of art after all. Renaming them to "destination triple" and "destination platform" arguably was a mistake. Since we had no tagged releases in the meantime, let's clean up the naming before it's too late.

`--experimental-cross-compilation-sdks-path` is renamed to `--experimental-destinations-path`, `sharedCCSDKDirectory` to `sharedDestinationsDirectory`. `destinationTriple` is renamed to `targetTriple`.