Add macOS version to Package.swift for newly generated packages
Previously if you created a new basic Swift package and added a
dependency, such as swift-syntax, that requires a minimum version
greater than macOS 10.13, you would have to set the platform version to
something above that in the Package.swift. With this change newly
generated packages on macOS automatically generate the current macOS
version as the minimum supported version, and if the developer wants a
lower version they can reduce it as usual in the Package.swift.