Add package metadata to observability scopes used by PackageContainers
PackageContainer is a protocol that is implemented by various types involved in SwiftPM package resolution.  They can emit messages but those messages don't get associated with the diagnostics.

There are currently three implementors of this protocol:  FileSystemPackageContainer, SourceControlPackageContainer, and RegistryPackageContainer.  All three just assign a given ObservabilityScope, and setting a package location metadata on that scope is a good way to make sure that all the diagnostics emitted to that scope get associated with the package.

This fix makes sure that diagnostics emitted through these scopes get properly associated with the package.  This is particularly useful for manifest-related diagnostics, which inherit the scope used by the various kinds of PackageContainer.

rdar://103229985