SwiftPM outputs raw diagnostics from SwiftDriver without trailing newlines
SwiftPM uses the Swift Driver's `-parseable-output` flag to get a sequence of length-prefixed JSON-encoded messages that it can read.  Unfortunately the Swift Driver doesn't JSON-encode its own diagnostics, leading to things like https://github.com/apple/swift-package-manager/issues/5968.

I filed https://github.com/apple/swift-package-manager/issues/5968 on the Swift Driver to get it to encode its JSON messages, but meanwhile, it turns out that we can fairly easily fix the fallback logic that SwiftPM uses when it emits raw output.  It was simply missing a newline.  It's safe to always add a newline to the raw driver output, since the logic that parses JSON splits by newlines, so we won't find any terminating newlines that shouldn't be on the string.

Note that we do not add newlines to the output coming from messages from the compiler frontend, since that output already has trailing newlines.  So in essence this change makes the `.unparsableOutput()` case the same as the cases of regular output, from a newline perspective.

rdar://103608636